Drag and drop favorite bug (#4280)
